refactor test constants
