+++ src/test/java/uk/gov/companieshouse/digitalcertifiedcopyprocessor/consumer/InvalidMessageRouterTest.java
@@ -13,7 +13,7 @@ import java.util.Map;
 import static org.hamcrest.CoreMatchers.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.when;
-import static uk.gov.companieshouse.digitalcertifiedcopyprocessor.consumer.ConsumerTestConstants.ITEM_ORDERED_CERTIFIED_COPY;
+import static uk.gov.companieshouse.digitalcertifiedcopyprocessor.util.Constants.ITEM_ORDERED_CERTIFIED_COPY;
 
 @ExtendWith(MockitoExtension.class)
 class InvalidMessageRouterTest {
